backend/user_service/requirements.txt
fastapi
uvicorn[standard]
asyncpg
psycopg2-binary
bcrypt
PyJWT
python-multipart
email-validator
